refactor: use modal
