+++ src/extension/app/constants.js
@@ -59,6 +59,7 @@ export const MODALS = {
   WAIT: 'wait',
   ERROR: 'error',
   DELETE: 'delete',
+  PLUGIN_LIST: 'plugin-list',
 };
 
 /**
@@ -125,6 +126,10 @@ export const ICONS = {
       <path d="M14.4424 18.499C14.1323 18.499 13.8218 18.4092 13.5469 18.229L11.1758 16.6719C10.4629 16.2031 9.49463 16.2031 8.78076 16.6719L6.41015 18.229C5.83788 18.605 5.11083 18.5869 4.55712 18.185C4.00341 17.7832 3.76269 17.0967 3.94286 16.4365L4.90331 12.9228C5.0039 12.5557 4.87694 12.1645 4.58007 11.9268L1.7373 9.64892C1.20312 9.2207 0.995108 8.52343 1.20654 7.87255C1.41845 7.22216 1.99658 6.78075 2.67968 6.74853L6.31786 6.57617C6.69823 6.55859 7.03075 6.31689 7.16552 5.96045L8.45409 2.55322C8.69579 1.91357 9.29393 1.50049 9.97802 1.5C10.6621 1.5 11.2607 1.91357 11.5029 2.55322L12.791 5.96045C12.9258 6.3169 13.2583 6.55859 13.6387 6.57617L17.2773 6.74853C17.9604 6.78076 18.5386 7.22216 18.7505 7.87304C18.9619 8.52392 18.7534 9.2207 18.2192 9.64892L15.3769 11.9268C15.0801 12.1645 14.9531 12.5557 15.0537 12.9224L16.0141 16.437C16.1943 17.0967 15.9531 17.7832 15.3999 18.1851C15.1123 18.394 14.7778 18.499 14.4424 18.499Z" fill="currentColor"/>
     </svg>`,
   ARROW: html`
+    <svg height="18" width="18" viewBox="0 0 22 22" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
+      <path d="M19.75,10.04h-15l5.97-5.97a.483.483,0,0,0,0-.7l-.35-.36a.513.513,0,0,0-.71,0L2.24,10.44a.513.513,0,0,0,0,.71l7.39,7.84a.513.513,0,0,0,.71,0l.35-.35a.513.513,0,0,0,0-.71L4.76,11.5H19.75a.25.25,0,0,0,.25-.25v-.96A.25.25,0,0,0,19.75,10.04Z"></path>
+    </svg>`,
+  RETURN: html`
     <svg height="18" width="18" viewBox="0 0 22 22" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
       <path d="M19.75,10.04h-15l5.97-5.97a.483.483,0,0,0,0-.7l-.35-.36a.513.513,0,0,0-.71,0L2.24,10.44a.513.513,0,0,0,0,.71l7.39,7.84a.513.513,0,0,0,.71,0l.35-.35a.513.513,0,0,0,0-.71L4.76,11.5H19.75a.25.25,0,0,0,.25-.25v-.96A.25.25,0,0,0,19.75,10.04Z"></path>
     </svg>`,
