+++ src/extension/index.js
@@ -42,6 +42,7 @@ import './app/components/plugin/palette-container.js';
 import './app/components/plugin/palette-dialog/palette-dialog-wrapper.js';
 import './app/components/plugin/palette-dialog/palette-dialog.js';
 import './app/components/plugin/env-switcher/env-switcher.js';
+import './app/components/plugin/plugin-list/plugin-list.js';
 import './app/components/plugin/login/login.js';
 import './app/components/dialog-wrapper/dialog-wrapper.js';
